Changed the data type of the jwtExpirationDate in JwtTokenProvider class to long to accommodate a lengthy expiry date

+++ src/main/java/com/springboot/blog/security/JwtTokenProvider.java
@@ -15,7 +15,7 @@ public class JwtTokenProvider {
     @Value("${app.jwt-secret}")
     private String jwtSecretKey;
     @Value("${app.jwt-expiration-milliseconds}")
-    private int jwtExpirationDate;
+    private Long jwtExpirationDate;
 
 
     //Generate token
